+++ lib/project_types/extension/messages/messages.rb
@@ -200,6 +200,9 @@ module Extension
               {{*}} You’re ready to start building {{green:%s}}!
             MESSAGE
           },
+          serve: {
+            unsupported: "shopify extension serve is not supported for theme app extensions",
+          },
         },
       },
     }