+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ------
 
+* [#1386](https://github.com/Shopify/shopify-cli/pull/1386): Update Theme-Check to 1.2
+
 Version 2.1.0
 -------------
 * [#1357](https://github.com/Shopify/shopify-cli/pull/1357): Update Theme-Check to 1.1