+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ------
+* [#1197](https://github.com/Shopify/shopify-app-cli/pull/1197): Add `create` command for PHP app projects.
 * [#1399](https://github.com/Shopify/shopify-cli/pull/1399): Fix error when running `shopify extension serve` in a theme app extension project
 
 Version 2.1.0